+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## [Unreleased]
 
+## [0.9.2] - 2020-12-22
+
+### Added
+ * Customizable threshold for percentage of capturing monitors
+
+### Fixed
+ * Capturing monitor percentage threshold
+ * Threshold event clearing
+ * Storage detection
+
 ## [0.9.1] - 2020-12-21
 
 ### Fixed
@@ -25,6 +35,7 @@
 ## 0.7.0 - 2019-03-17
  * Alpha release
 
-[Unreleased]: https://github.com/daviswr/ZenPacks.daviswr.ZoneMinder/compare/0.9.1...HEAD
-[0.9.1]: https://github.com/daviswr/ZenPacks.daviswr.ZoneMinder/compare/0.9.1...0.9.1
+[Unreleased]: https://github.com/daviswr/ZenPacks.daviswr.ZoneMinder/compare/0.9.2...HEAD
+[0.9.2]: https://github.com/daviswr/ZenPacks.daviswr.ZoneMinder/compare/0.9.1...0.9.2
+[0.9.1]: https://github.com/daviswr/ZenPacks.daviswr.ZoneMinder/compare/0.9.0...0.9.1
 [0.9.0]: https://github.com/daviswr/ZenPacks.daviswr.ZoneMinder/compare/0.7.0...0.9.0